==> scripts/1_import_data_toxo_neo_horm.R <==
###############################################################################
##############       Associations of Toxoplasma gondii and       ##############
##############       Neospora caninum with hormone levels        ##############
##############                  1. Data Import                   ##############
##############                 By: Zach Laubach                  ##############
##############               created: 19 Sept 2020               ##############
##############             last updated: 21 Sept 2020            ##############
###############################################################################


  ### PURPOSE: 
  
  
  # Code Blocks
    # 1: Configure workspace
    # 2: Import data 
    # 3: Export data files
  



###############################################################################
##############             1.  Configure workspace               ##############
###############################################################################

  ### 1.1 Global options
    ## a) clear global environment
      rm(list = ls())

    ## b) prevent R from automatically reading charater strins as factors
      options(stringsAsFactors = FALSE)
  

  ### 1.2 Install and load CRAN packages   
    ## a) Data Manipulation and Descriptive Stats Packages
      # Check for tidyverse and install if not already installed
        if (!'tidyverse' %in% installed.packages()[,1]){
          install.packages ('tidyyverse')
        }
      # load tidyverse packages
        library ('tidyverse')
      
        
      # Check for here and install if not already installed
        if (!'here' %in% installed.packages()[,1]){
          install.packages ('here')
        }
      # load here packages
        library ('here')
      
      
  ### 1.3 Install and load Mara Hyena Project packages 
    ## a) Load the Mara Hyena Project data files from github
      # Check for devtools and install if not already installed
      if(!'devtools' %in% row.names(installed.packages())){
        install.packages('devtools')
      }
    ## b) Use devtools to install hyeanadata package from the MaraHyenaProject
      # on github
      # devtools::install_github('MaraHyenaProject/hyenadata',
      #                 auth_token = "###")
      #******** NOTE: this requires permission and an authentication key. Please
      #********       send request to zchlaubach@gmail.com
      
      # load hyenadata package
      library('hyenadata')

        
  ### 1.4 Get Version and Session Info
    R.Version()
    sessionInfo()
    
    # Developed in:   
    # R version 4.0.0 (2020-04-24)
    # Platform: x86_64-apple-darwin17.0 (64-bit)
    # Running under: macOS Mojave 10.14.6
    
  
  ### 1.5 Set working directory 
    setwd(here())
  
  
  ### 1.6 Set file paths for data importing and exporting
    ## a) The path to sample, normalized RNA expression, and transformed behavior data
      project_data_path <- paste0(here('data/'))
     
  
  
      
###############################################################################
##############                  2. Import data                   ##############
###############################################################################    
      
  ### 2.1 Import sample data files
    ## a) Import nesop_toxo_data.
      neosp_toxo_data <- read_csv(paste0(project_data_path,
                                  'neosp_toxo_data.csv'))
      
  ### 2.2 Import Access Fisi data files
    ## a) Import Access data backend from Mara Hyena Project data package
      # Use hyenadata package 'load_all_tables' function to load all tables
      hyenadata::load_all_tables() 
      
    ## b) Rename data tables
      fecal_horm <- tblFecalHormones
      fecal_repos <- tblFecalRepository
      repro_state <- tblReproStates
 
      

###############################################################################
##############                3. Export data files               ##############
###############################################################################
      
  ### 3.1 Export data to an RData file     
    ## a) Save and export raw data tables 
      # Files are saved in the 'data' folder in the working directory as an
      # RData file.
      save(file = paste0(project_data_path, 
                         'raw_data_neosp_toxo_hormone.RData'), 
           list = c('neosp_toxo_data', 'fecal_horm', 
                    'fecal_repos', 'repro_state'))
